Updated VSliderFloat sol binding

+++ sol_ImGui.h
@@ -2430,7 +2430,7 @@ namespace sol_ImGui
 		ImGui.set_function("VSliderFloat"					, sol::overload(
 																sol::resolve<std::tuple<float, bool>(const std::string&, float, float, float, float, float)>(VSliderFloat),
 																sol::resolve<std::tuple<float, bool>(const std::string&, float, float, float, float, float, const std::string&)>(VSliderFloat),
-																sol::resolve<std::tuple<float, bool>(const std::string&, float, float, float, float, float, const std::string&, float)>(VSliderFloat)
+																sol::resolve<std::tuple<float, bool>(const std::string&, float, float, float, float, float, const std::string&, int)>(VSliderFloat)
 															));
 		ImGui.set_function("VSliderInt"						, sol::overload(
 																sol::resolve<std::tuple<int, bool>(const std::string&, float, float, int, int, int)>(VSliderInt),
